Commit END

+++ src/main/java/com/fanfiction/service/AdminService.java
@@ -1,11 +1,8 @@
 package com.fanfiction.service;
 
-import com.fanfiction.models.ERole;
-import com.fanfiction.models.User;
+import com.fanfiction.models.*;
 import com.fanfiction.DTO.UserJwtDTO;
-import com.fanfiction.models.Role;
-import com.fanfiction.repository.RoleRepository;
-import com.fanfiction.repository.UserRepository;
+import com.fanfiction.repository.*;
 import com.fanfiction.security.jwt.JwtUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,23 +17,38 @@ import java.util.stream.Collectors;
 public class AdminService {
 
     @Autowired
+    private ChapterRepository chapterRepository;
+
+    @Autowired
     private UserRepository userRepository;
 
     @Autowired
     private RoleRepository roleRepository;
 
     @Autowired
+    private CommentRepository commentRepository;
+
+    @Autowired
+    private CompositionRepository compositionRepository;
+
+    @Autowired
     private JwtUtils jwtUtils;
 
-    public List<User> getAllUsers(){
+    public List<User> getAllUsers() {
         return userRepository.findAll();
     }
 
-    public void deleteUser(Long userId){
-        userRepository.deleteById(userId);
+    public void deleteUser(Long userId) {
+        if (!(userRepository.findById(userId).get().getRoles().contains(roleRepository.findByName(ERole.ROLE_ADMIN).get()))) {
+            List<Composition> compositions = compositionRepository.findCompositionsByAuthorId(userId);
+            compositions.forEach(composition -> chapterRepository.deleteAllByComposition(composition));
+            compositionRepository.deleteAll(compositions);
+            commentRepository.deleteAll(commentRepository.findAllByCommentUserId(userId));
+            userRepository.deleteById(userId);
+        }
     }
 
-    public void setRole(Long userId, ERole role){
+    public void setRole(Long userId, ERole role) {
         User user = userRepository.findById(userId).get();
         Set<Role> roles = new HashSet<>();
         roles.add(roleRepository.findByName(role).get());
@@ -44,7 +56,7 @@ public class AdminService {
         userRepository.save(user);
     }
 
-    public UserJwtDTO getUser(Long userId){
+    public UserJwtDTO getUser(Long userId) {
         User user = userRepository.findById(userId).get();
         return new UserJwtDTO(jwtUtils.generateJwtTokenByUsername(user.getUsername()),
                 userId,
